pico: merge st7789 files
It's not the pimoroni-pico driver any more, there's no real reason to keep it seperate
